Move concurrency state and config to a separate module
To avoid circular imports.
