+++ scripts/tests/chiptest/concurrency/process.py
@@ -13,169 +13,20 @@
 # limitations under the License.
 
 import contextlib
-import dataclasses
-import enum
-import functools
 import logging
-import multiprocessing
 import os
 import signal
-import threading
-import traceback
 from abc import ABC, abstractmethod
-from collections.abc import Callable, Iterator
 from multiprocessing.context import SpawnContext
-from multiprocessing.managers import SyncManager, ValueProxy
-from types import TracebackType
-from typing import ClassVar, Concatenate, Generic, ParamSpec, Protocol, Self, TypeVar
+from multiprocessing.managers import SyncManager
+from typing import Generic, TypeVar
 
-from chiptest.concurrency.work_queue import CancellableQueue, EndOfQueue, QueueCancelled, wait_for_mp_managed
-from chiptest.log_config import LogConfig
+from chiptest.concurrency.config import ProcessConfig
+from chiptest.concurrency.context import with_annotated_exception
+from chiptest.concurrency.state import ProcessPhase, ProcessState
+from chiptest.concurrency.work_queue import CancellableQueue, EndOfQueue, QueueCancelled
 
 log = logging.getLogger(__name__)
-
-
-class WithName(Protocol):
-    @property
-    def name(self) -> str: ...
-
-
-S = TypeVar("S", bound="WithName")
-P = ParamSpec("P")
-R = TypeVar("R")
-
-
-def with_annotated_exception(fn: Callable[Concatenate[S, P], R]) -> Callable[Concatenate[S, P], R]:
-    """Decorator to enrich exceptions from thread and process methods with their names for easier debugging."""
-
-    @functools.wraps(fn)
-    def wrapper(self: S, *args: P.args, **kwargs: P.kwargs) -> R:
-        try:
-            return fn(self, *args, **kwargs)
-        except BaseException as e:
-            kind = ("thread" if isinstance(self, threading.Thread) else
-                    "process" if isinstance(self, (multiprocessing.Process, WrappedProcess)) else
-                    self.__class__.__name__)
-
-            e.add_note(f"Exception in {kind} {self.name}")
-            raise
-
-    return wrapper
-
-
-@dataclasses.dataclass
-class ProcessConfig:
-    """Process configuration template."""
-
-    DEFAULT_START_TIMEOUT_SEC: ClassVar[float] = 4.0
-    DEFAULT_STOP_TIMEOUT_SEC: ClassVar[float] = 4.0
-    DEFAULT_TERMINATION_TIMEOUT_SEC: ClassVar[float] = 2.0
-
-    id: int = 0
-    name: str = "Process{id}"
-    log_config: LogConfig = dataclasses.field(default_factory=LogConfig)
-    start_timeout_sec: float = DEFAULT_START_TIMEOUT_SEC
-    stop_timeout_sec: float = DEFAULT_STOP_TIMEOUT_SEC
-    termination_timeout_sec: float = DEFAULT_TERMINATION_TIMEOUT_SEC
-
-    def with_formatted_name(self) -> Self:
-        """Format the name using the id and process hierarchy. Return a new instance with the formatted name."""
-        name = self.name.format(id=self.id)
-
-        # Indicate process hierarchy in the name so that it's easier to understand the dependencies in the logs.
-        if (parent_proc_name := multiprocessing.current_process().name) != "MainProcess":
-            name = f"{parent_proc_name}/{name}"
-
-        return dataclasses.replace(self, name=name)
-
-
-class ProcessPhase(enum.IntEnum):
-    """Phase of the process lifecycle."""
-
-    NOT_STARTED = enum.auto()
-    """The process has not started yet."""
-
-    UNINITIALIZED = enum.auto()
-    """The process has started but has not finished initialization yet."""
-
-    READY = enum.auto()
-    """The process has finished initialization and is ready to work. Can cycle between READY and WORKING phases multiple times."""
-
-    WORKING = enum.auto()
-    """The process is working on a task. Optional and can be used to distinguish between idle and busy processes."""
-
-    CLOSED = enum.auto()
-    """The process has finished all work and is closed (exited). Mind that it could have been closed due to an exception."""
-
-
-class ProcessState:
-    """
-    Shared process state for synchronized lifecycle coordination and error propagation between the parent and child processes.
-
-    Can be used as a context manager to synchronize access to the state. Needed only if accessing multiple fields atomically,
-    otherwise individual fields are synchronized internally.
-    """
-
-    def __init__(self, mp_manager: SyncManager, config: ProcessConfig) -> None:
-        self._config = config
-        self._phase: ValueProxy[ProcessPhase] = mp_manager.Value(object, ProcessPhase.NOT_STARTED)
-        self._exception: ValueProxy[BaseException | None] = mp_manager.Value(object, None)
-        self._state_changed = mp_manager.Condition()
-
-    @property
-    def phase(self) -> ProcessPhase:
-        return self._phase.get()
-
-    @phase.setter
-    def phase(self, value: ProcessPhase) -> None:
-        with self._state_changed:
-            self._phase.set(value)
-            self._state_changed.notify_all()
-
-    @property
-    def exception(self) -> BaseException | None:
-        return self._exception.get()
-
-    @exception.setter
-    def exception(self, value: BaseException | None) -> None:
-        with self._state_changed:
-            if isinstance(value, BaseException):
-                value.add_note(f"Exception in process {self._config.name}")
-
-                # Traceback is not automatically propagated across process boundaries, so we need to add it to the exception
-                # manually as a note.
-                value.add_note("".join(traceback.format_exception(value)))
-            self._exception.set(value)
-            self._state_changed.notify_all()
-
-    def __enter__(self) -> bool:
-        return self._state_changed.__enter__()
-
-    def __exit__(self, exc_type: type[BaseException] | None, exc_val: BaseException | None, exc_tb: TracebackType | None):
-        return self._state_changed.__exit__(exc_type, exc_val, exc_tb)
-
-    def wait_for(self, predicate: Callable[[ProcessPhase, BaseException | None], bool], timeout: float | None = None):
-        """
-        Wait until the predicate returns True for the current phase and/or exception.
-
-        Timeout as for `Condition.wait_for()`, i.e. it can be a positive float for maximum wait time, or None to wait indefinitely.
-
-        Returns the result of the predicate.
-
-        We need the wait_for_mp_managed wrapper because otherwise we wouldn't be able to catch a KeyboardInterrupt for the condition
-        which is managed by multiprocessing.Manager, as the manager process explicitly ignores SIGINT.
-        """
-        with self._state_changed:
-            return wait_for_mp_managed(self._state_changed, lambda: predicate(self._phase.get(), self._exception.get()), timeout)
-
-    @contextlib.contextmanager
-    def working_context(self) -> Iterator[None]:
-        """Context manager to set the process phase to WORKING for the duration of the context."""
-        self.phase = ProcessPhase.WORKING
-        try:
-            yield
-        finally:
-            self.phase = ProcessPhase.READY
 
 
 WorkerConfigT = TypeVar("WorkerConfigT", bound=ProcessConfig)
